wxGUI: icon size is float, not int (https://trac.osgeo.org/grass/ticket/1035)
git-svn-id: https://svn.osgeo.org/grass/grass/branches/releasebranch_6_4@52917 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ gui/wxpython/mapdisp/frame.py
@@ -879,7 +879,7 @@ class MapFrame(MapFrameBase):
             if '=' in p:
                 parg,pval = p.split('=')
                 if parg == 'icon': icon = pval
-                elif parg == 'size': size = int(pval)
+                elif parg == 'size': size = float(pval)
 
         pattern = ["d.vect",
                    "map=%s" % name,
